Change link order in makefile
When compiling with "mix deps.compile" (on Ubuntu 18.04 with GCC 7.5.0)
an "unused symbol" error in the NIF library occured.
Examination of the .so files showed that libportmidi
was not linked into the shared library.
It seems like --as-needed is a default linker option in Ubuntu,
which leads to portmidi not beeing linked at all in this case.
Changing the linking order solves this issue.